requirements.txt
# GPT Researcher dependencies
gpt-researcher>=0.14.0
python-dotenv

# MCP dependencies
fastmcp>=2.8.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0

# Utility dependencies
loguru>=0.7.0

# Gemini dependencies
google-genai>=1.0.0

# Testing dependencies
httpx>=0.24.0
sseclient-py>=1.7.0
